compose add favorites - lvl1 complete
