+++ app/src/main/java/com/example/playlistmaker/presentation/favorites/activity/FavoritesFragment.kt
@@ -4,6 +4,11 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.ComposeView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
@@ -14,26 +19,31 @@ import com.example.playlistmaker.presentation.favorites.models.FavoritesScreenSt
 import com.example.playlistmaker.presentation.favorites.view_model.FavoritesViewModel
 import com.example.playlistmaker.presentation.player.activity.PlayerFragment
 import com.example.playlistmaker.presentation.search.activity.SearchAdapter
+import com.example.playlistmaker.ui.library.Favorites
+import com.example.playlistmaker.ui.search.Search
+import com.example.playlistmaker.ui.theme.PlaylistMakerTheme
 import com.example.playlistmaker.util.debounce
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class FavoritesFragment : Fragment() {
 
-    private var _binding: FragmentFavouritesBinding? = null
-    private val binding get() = _binding!!
-
     private val favoritesViewModel: FavoritesViewModel by viewModel()
 
     private lateinit var clickListenerDebounce: (Track) -> Unit
 
-    private val favoritesAdapter by lazy {
-        SearchAdapter { trackItem -> clickListenerDebounce(trackItem) }
-    }
-
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View {
-        _binding = FragmentFavouritesBinding.inflate(inflater, container, false)
-        return binding.root
+        return ComposeView(requireContext()).apply {
+            setContent {
+                PlaylistMakerTheme {
+                    Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
+                        Favorites(modifier = Modifier.padding(innerPadding), favoritesViewModel) {
+                                trackItem -> clickListenerDebounce(trackItem)
+                        }
+                    }
+                }
+            }
+        }
     }
 
     override fun onResume() {
@@ -43,7 +53,6 @@ class FavoritesFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        binding.favoritesRecyclerView.adapter = favoritesAdapter
 
         clickListenerDebounce = debounce<Track>(
             CLICK_DEBOUNCE_DELAY,
@@ -51,43 +60,12 @@ class FavoritesFragment : Fragment() {
                 trackItem -> trackListClickListener(trackItem)
         }
 
-        favoritesViewModel.stateLiveData.observe(viewLifecycleOwner) { state ->
-            render(state)
-        }
-
-    }
-
-    private fun render(state: FavoritesScreenState) {
-        when (state) {
-            is FavoritesScreenState.Empty -> showEmpty()
-            is FavoritesScreenState.Content -> showContent(state.trackList)
-        }
-    }
-
-    private fun showEmpty() = with(binding) {
-        favoritesRecyclerView.visibility = View.GONE
-        placeholderTextView.visibility = View.VISIBLE
-        placeholderImageView.visibility = View.VISIBLE
-    }
-
-    private fun showContent(trackList: List<Track>) {
-        favoritesAdapter.items = trackList
-        with(binding) {
-            favoritesRecyclerView.visibility = View.VISIBLE
-            placeholderTextView.visibility = View.GONE
-            placeholderImageView.visibility = View.GONE
-        }
     }
 
     private fun trackListClickListener(trackItem: Track) {
         findNavController().navigate(
             R.id.action_libraryFragment_to_playerFragment,
             PlayerFragment.createArgs(trackItem))
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
     }
 
     companion object {
